+++ .gitignore
@@ -1,3 +1,1 @@
 node_modules/
-
-local-data/
